destination-redshift: no more standard inserts
